final-pjt/final_pjt_back/SD_api/serializers.py
from rest_framework import serializers
from .models import SavingProduct, DepositProduct

class SavingProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = SavingProduct
        fields = '__all__'

class DepositProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = DepositProduct
        fields = '__all__'
